% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup Data}
\format{
A data frame with 22 observations with 4 columns:
\describe{
\item{WorldCup}{Country name and year of occurrence}
\item{Matches}{Number of matches played in the World Cup}
\item{Totalattendance}{Total attendance for all matches in the World Cup}
\item{Averageattendance}{Average attendance per match in the World Cup}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
A dataset containing FIFA World Cup attendance from 1930 to 2022.
}
\keyword{datasets}
