% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{UN World Country Population Data from 1950 to 2020}
\format{
A data frame with 16685 observations with 3 columns:
\describe{
\item{Country Name}{Name of the country}
\item{Year}{Year of the population estimate}
\item{Population Estimates}{Estimated population of the country}
}
}
\source{
World_Population.xlsx file
}
\usage{
WorldPopulation
}
\description{
A dataset containing population data for all UN countries from 1950 to 2020.
}
\keyword{datasets}
